Broke files up to better manage this project
