+++ src/ARVRInterface.h
@@ -1,0 +1,13 @@
+////////////////////////////////////////////////////////////////////////////////////////////////
+// Our main ARVRInterface code for our OpenHMD GDNative module
+
+#ifndef OVR_ARVR_INTERFACE_H
+#define OVR_ARVR_INTERFACE_H
+
+#include "GodotCalls.h"
+#include "openhmd_data.h"
+#include "openhmd_shader.h"
+
+extern const godot_arvr_interface_gdnative interface_struct;
+
+#endif /* !OVR_ARVR_INTERFACE_H */
